tasks/common/common -> tasks/common/{fetch,misc}
