+++ qgreenland/tasks/layers/arctic_dem.py
@@ -1,6 +1,6 @@
 import os
 
-from qgreenland.tasks.common.common import FetchDataFile
+from qgreenland.tasks.common.fetch import FetchDataFile
 from qgreenland.tasks.common.raster import (BuildRasterOverviews,
                                             ReprojectRaster,
                                             SubsetRaster)
